requirements.txt
discord.py
python-dotenv
requests
beautifulsoup4
zenrows
asyncio